integrate parse_result_live.py in parse_result.py
